# docs/swagger.yaml
definitions:
...
  controller.createAccountRequest:
    properties:
      currency:
        type: string
      phone_number:
        type: string
    type: object
...
  controller.getAccountByCurrencyRequest:
    properties:
      currency:
        type: string
    required:
    - currency
    type: object
...
  models.Account:
    properties:
      active:
        type: boolean
      balance:
        type: integer
      createdAt:
        type: string
      currency:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      phoneNumber:
        type: string
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
...
  models.UpdateAccount:
    properties:
      balance:
        type: integer
      currency:
        type: string
      id:
        type: integer
      phone_number:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /accounts:
    patch:
      consumes:
      - application/json
      description: Updates an account's phone number, balance, or currency for the
        authenticated user
      parameters:
      - description: Account update data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Invalid input, account not found, or invalid currency
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an existing account
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Creates a new account for the authenticated user with the provided
        currency and phone number
      parameters:
      - description: Account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/controller.createAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Invalid input, user not found, account already exists, or invalid
            currency
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new account
      tags:
      - accounts
  /accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an account for the authenticated user by account ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Account deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input, account not found, or non-zero balance
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an existing account
      tags:
      - accounts
    get:
      consumes:
      - application/json
      description: Retrieves an account by its ID for the authenticated user
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Invalid input or account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an account by ID
      tags:
      - accounts
  /accounts/{id}/balance:
    get:
      consumes:
      - application/json
      description: Retrieves an account balance by its ID for the authenticated user
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Invalid input or account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an account balance by ID
      tags:
      - accounts
  /accounts/currency:
    get:
      consumes:
      - application/json
      description: Retrieves all accounts with the specified currency for the authenticated
        user
      parameters:
      - description: Currency
        in: body
        name: currency
        required: true
        schema:
          $ref: '#/definitions/controller.getAccountByCurrencyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Account'
            type: array
        "400":
          description: Invalid input or invalid currency
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get accounts by currency
      tags:
      - accounts
  /accounts/users/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves an account by its ID for the authenticated user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Invalid input or account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an account by user ID
      tags:
      - accounts
...
swagger: "2.0"